% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_values.R
\name{format_values}
\alias{format_values}
\title{Format numeric values}
\usage{
format_values(x, digits = getOption("surveyvisualize.format.digits"),
  percentage = getOption("surveyvisualize.format.percentage"),
  scientific = getOption("surveyvisualize.format.scientific"),
  multiplier = getOption("surveyvisualize.format.multiplier"))
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{integer. Number of decimal places to use.}

\item{percentage}{logical. Treat proportions as percentages?}

\item{scientific}{logical. Use scientific notation for number formatting?}

\item{multiplier}{numeric. A multiplier to use for numeric value display (i.e. For "In Thousands", use multiplier = 1000)}
}
\value{
Formatted character vector, the same length as x.
}
\description{
Consistent formatting options across all visualization functions.
}
