% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chart.R
\name{make_chart}
\alias{make_chart}
\title{Make Interactive Charts}
\usage{
make_chart(
  tbl,
  x = NULL,
  y = NULL,
  fill = NULL,
  facet = NULL,
  interactive = TRUE,
  order = FALSE,
  flip = FALSE,
  palette = "Set3",
  variable_labels = TRUE,
  label_options = list(wrap = 35, trunc = 100),
  legend = TRUE,
  confidence = 0.95,
  use_girafe = TRUE,
  girafe_options = list(),
  ...
)
}
\arguments{
\item{tbl}{Analysis table (data.table object).}

\item{x}{character. Variable to be displayed along x-axis.}

\item{y}{character. Response variable.}

\item{fill}{character. Variable to be categorized using interior coloring.}

\item{facet}{logical. Variable to be categorized by faceted grid.}

\item{interactive}{logical. Print as girafe object instead of ggplot object?}

\item{order}{logical. Order by values instead of natural order?}

\item{flip}{logical. Flip x and y axes?}

\item{palette}{character. A color pallete to be used for the fill variable. See pallete argument here \link[ggplot2]{scale_fill_brewer}}

\item{legend}{logical. Include a legend?}

\item{confidence}{Confidence level for margin of error calculation. Defaults to 0.90. Set to NULL for standard error.}

\item{girafe_options}{Optional list of options to pass to \link[ggiraph]{girafe}.}

\item{...}{Optional formatting arguments. See \link[surveyvisualize]{format_values}.}

\item{flat_print}{logical. Print as ggplot object instead of girafe object?}
}
\description{
Make interactive bar charts utilizing \link[ggiraph]{girafe}.
}
