% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Make Interactive Maps}
\usage{
make_map(tbl, tbl_geo_id, geo_layer,
  geo_layer_id = guess_geo_layer_id(geo_layer),
  geo_layer_name = guess_geo_layer_name(geo_layer), tbl2 = NULL,
  confidence = 0.95, use_ggiraph = TRUE, ...)
}
\arguments{
\item{tbl}{Analysis table (data.table object).}

\item{tbl_geo_id}{Name of the geography variable in tbl.}

\item{geo_layer}{A spatial feature object or path to shapefile. The geography layer to display.}

\item{geo_layer_id}{The unique id of the spatial feature object. By default, guesses the id using guess_geo_layer_id.}

\item{geo_layer_name}{The name field in the spatial feature object. By default, guesses the name using guess_geo_layer_name.}

\item{tbl2}{Optional second table. Requires same geography group variable as tbl. tbl2 gets passed to \link[surveyvisualize]{make_chart} and output as an interactive tooltip over the matching tbl geography.}

\item{confidence}{Confidence level for margin of error calculation. Defaults to 0.90. Set to NULL for standard error.}

\item{...}{Optional formatting arguments. See \link[surveyvisualize]{format_values}.}
}
\value{
ggiraph/htmlwidget class object
}
\description{
Make interactive maps utilizing \link[ggiraph]{ggiraph}.
}
