% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Make HTML Table}
\usage{
make_table(tbl, row_vars = NULL, col_vars = NULL, confidence = 0.95,
  variable_labels = TRUE, use_viewer = TRUE, ...)
}
\arguments{
\item{tbl}{Analysis table (data.table object).}

\item{row_vars}{Group variables to be represented in the row position.}

\item{col_vars}{Group variables to be represented in the column position.}

\item{confidence}{Confidence level for margin of error calculation. Defaults to 0.90. Set to NULL for standard error.}

\item{...}{Optional formatting arguments. See \link[surveyvisualize]{format_values}.}

\item{title}{Title of the table}
}
\value{
A kableExtra object.
}
\description{
Make formatted HTML tables utilizing \link[kableExtra]{kableExtra-package}
}
