% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Make HTML Table}
\usage{
make_table(
  tbl,
  row_vars = NULL,
  col_vars = NULL,
  confidence = 0.95,
  variable_labels = TRUE,
  use_viewer = TRUE,
  stat_vars = c("Estimate", "SE", "Survey", "N"),
  stat_vars_output_names = list(Estimate = "Estimate", SE = "SE", Survey = "Survey", N =
    "N"),
  ...
)
}
\arguments{
\item{tbl}{Analysis table (data.table object).}

\item{row_vars}{Group variables to be represented in the row position.}

\item{col_vars}{Group variables to be represented in the column position.}

\item{confidence}{Confidence level for margin of error calculation. Defaults to 0.90. Set to NULL for standard error.}

\item{variable_labels}{logical. Use labels for variable names?}

\item{use_viewer}{logical. If interactive, render markdown table as html to be displayed in viewer pane?}

\item{stat_vars}{Select the statistic output columns to include or change or rearrange their default order.}

\item{...}{Optional formatting arguments. See \link[surveyvisualize]{format_values}.}
}
\value{
An html, pdf, or docx table.
}
\description{
Make formatted tables.
}
